+++ packages/mint-components/src/components/sqm-portal-reset-password/sqm-portal-reset-password.tsx
@@ -26,6 +26,35 @@ export class PortalResetPassword {
   @Prop()
   confirmPassword: boolean = false;
 
+  /**
+   * @uiName Password reset header text
+   */
+  @Prop() resetPasswordHeader: string = "Reset your password";
+
+  /**
+   * Displayed after a successful password reset
+   *
+   * @uiName Password reset header text
+   */
+  @Prop() passwordResetHeader: string = "Password reset";
+
+  /**
+   * @uiName Password reset button text
+   */
+  @Prop() resetPasswordButtonText: string = "Reset Password";
+
+  /**
+   * Displayed after a successful password reset
+   *
+   * @uiName Continue button text
+   */
+  @Prop() continueButtonText: string = "Continue";
+
+  /**
+   * @uiName Confirm password field label
+   */
+  @Prop() confirmPasswordFieldLabel: string = "Confirm Password";
+
   /** @undocumented */
   @Prop() demoData?: DemoData<PortalResetPasswordViewProps>;
 
@@ -54,6 +83,13 @@ function usePortalResetPasswordDemo(
         confirmPassword: true,
         oobCodeValidating: false,
         oobCodeValid: true,
+        content: {
+          passwordResetHeader: "Password reset",
+          resetPasswordHeader: "Reset your password",
+          continueButtonText: "Continue",
+          resetPasswordButtonText: "Reset Password",
+          confirmPasswordFieldLabel: "Confirm Password",
+        },
       },
       callbacks: {
         submit: async (_event: any) => {