+++ packages/mint-components/src/stories/PortalResetPassword.stories.tsx
@@ -20,6 +20,13 @@ const defaultProps: PortalResetPasswordViewProps = {
     confirmPassword: true,
     oobCodeValidating: false,
     oobCodeValid: true,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,
@@ -36,6 +43,13 @@ const defaultPropsNoConfirm: PortalResetPasswordViewProps = {
     confirmPassword: false,
     oobCodeValidating: false,
     oobCodeValid: true,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,
@@ -52,6 +66,13 @@ const errorProps: PortalResetPasswordViewProps = {
     confirmPassword: true,
     oobCodeValidating: false,
     oobCodeValid: true,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,
@@ -68,6 +89,13 @@ const loadingProps: PortalResetPasswordViewProps = {
     confirmPassword: true,
     oobCodeValidating: false,
     oobCodeValid: true,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,
@@ -84,6 +112,13 @@ const successProps: PortalResetPasswordViewProps = {
     confirmPassword: true,
     oobCodeValidating: false,
     oobCodeValid: true,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,
@@ -100,6 +135,13 @@ const validatingCodeProps: PortalResetPasswordViewProps = {
     confirmPassword: true,
     oobCodeValidating: false,
     oobCodeValid: false,
+    content: {
+      passwordResetHeader: "Password reset",
+      resetPasswordHeader: "Reset your password",
+      continueButtonText: "Continue",
+      resetPasswordButtonText: "Reset Password",
+      confirmPasswordFieldLabel: "Confirm Password",
+    },
   },
   callbacks: {
     submit: async (e) => await e,